fix tests and add pid tags
The tests now rely on patching the key system calls that produce the data that
will be recorded. This reduces duplication and allows the tests to exercise the
methods in the collector.

A `pid` tag has been added to all of the metrics, because there are situations
where this collector will be launched within each process in a forking model of
execution.